feat: integrate with testlogger.
