+++ CHANGELOG.md
@@ -1,0 +1,8 @@
+# Changelog
+
+## Unreleased (v3.0.x)
+
+* Refactor to support [core testlogger][]
+* Compatibility: minimum framework is netstandard1.5 and TestPlatform 15.5.0
+
+[core testlogger]: https://github.com/spekt/testlogger
