+++ scripts/dependencies.props
@@ -4,10 +4,10 @@
     <MSTestVersion>1.3.2</MSTestVersion>
     <NETTestSdkVersion>15.7.2</NETTestSdkVersion>
     <MoqVersion>4.9.0</MoqVersion>
-    <TestLoggerVersion>3.0.24</TestLoggerVersion>
+    <TestLoggerVersion>3.0.26</TestLoggerVersion>
 
     <!-- Test Assets use the minimum supported versions -->
-    <NETTestSdkMinimumVersion>15.0.0</NETTestSdkMinimumVersion>
+    <NETTestSdkMinimumVersion>15.5.0</NETTestSdkMinimumVersion>
     <XunitVersion>2.3.1</XunitVersion>
     <XunitTestAdapterVersion>2.3.1</XunitTestAdapterVersion>
