+++ scripts/version.props
@@ -1,6 +1,6 @@
 <Project>
   <PropertyGroup>
-    <PackageVersion Condition="'$(PackageVersion)' == ''">2.1.0-dev</PackageVersion>
+    <PackageVersion Condition="'$(PackageVersion)' == ''">3.0.0-dev</PackageVersion>
   </PropertyGroup>
 </Project>
 <!-- vi: set ft=xml: -->
