+++ src/Xunit.Xml.TestLogger.TestAdapter/Xunit.Xml.TestLogger.TestAdapter.csproj
@@ -19,6 +19,9 @@
     <Compile Include="..\Xunit.Xml.TestLogger\XunitXmlTestLogger.cs">
       <Link>XunitXmlTestLogger.cs</Link>
     </Compile>
+    <Compile Include="..\Xunit.Xml.TestLogger\XunitXmlSerializer.cs">
+      <Link>XunitXmlSerializer.cs</Link>
+    </Compile>
     <Compile Include="..\Xunit.Xml.TestLogger\TestResultInfo.cs">
       <Link>TestResultInfo.cs</Link>
     </Compile>
