+++ src/package/XunitXml.TestLogger.nuspec
@@ -18,9 +18,9 @@
     <file src="packageIcon.png" target="\" />
 
     <!-- Net Core -->
-    <file src="XunitXml.TestLogger.props" target="build\netstandard1.0\" />
+    <file src="XunitXml.TestLogger.props" target="build\netstandard1.5\" />
 
     <!-- Desktop -->
-    <file src="XunitXml.TestLogger.props" target="build\net45\" />
+    <file src="XunitXml.TestLogger.props" target="build\net46\" />
   </files>
 </package>
